requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
PyYAML==6.0.2
openpyxl==3.1.5
dnspython==2.7.0
tldextract==5.1.2
mangum==0.17.0
python-multipart==0.0.9
